Rename net/http patch file and require net/http
