+++ lib/mini_profiler/patches.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
 require 'mini_profiler/patches/sql_patches'
-require 'mini_profiler/patches/net_patches'
+require 'mini_profiler/patches/net/http'
